+++ src/Service/DynamoDb/src/ValueObject/CreateGlobalSecondaryIndexAction.php
@@ -44,11 +44,20 @@ final class CreateGlobalSecondaryIndexAction
     private $provisionedThroughput;
 
     /**
+     * The maximum number of read and write units for the global secondary index being created. If you use this parameter,
+     * you must specify `MaxReadRequestUnits`, `MaxWriteRequestUnits`, or both.
+     *
+     * @var OnDemandThroughput|null
+     */
+    private $onDemandThroughput;
+
+    /**
      * @param array{
      *   IndexName: string,
      *   KeySchema: array<KeySchemaElement|array>,
      *   Projection: Projection|array,
      *   ProvisionedThroughput?: null|ProvisionedThroughput|array,
+     *   OnDemandThroughput?: null|OnDemandThroughput|array,
      * } $input
      */
     public function __construct(array $input)
@@ -57,6 +66,7 @@ final class CreateGlobalSecondaryIndexAction
         $this->keySchema = isset($input['KeySchema']) ? array_map([KeySchemaElement::class, 'create'], $input['KeySchema']) : $this->throwException(new InvalidArgument('Missing required field "KeySchema".'));
         $this->projection = isset($input['Projection']) ? Projection::create($input['Projection']) : $this->throwException(new InvalidArgument('Missing required field "Projection".'));
         $this->provisionedThroughput = isset($input['ProvisionedThroughput']) ? ProvisionedThroughput::create($input['ProvisionedThroughput']) : null;
+        $this->onDemandThroughput = isset($input['OnDemandThroughput']) ? OnDemandThroughput::create($input['OnDemandThroughput']) : null;
     }
 
     /**
@@ -65,6 +75,7 @@ final class CreateGlobalSecondaryIndexAction
      *   KeySchema: array<KeySchemaElement|array>,
      *   Projection: Projection|array,
      *   ProvisionedThroughput?: null|ProvisionedThroughput|array,
+     *   OnDemandThroughput?: null|OnDemandThroughput|array,
      * }|CreateGlobalSecondaryIndexAction $input
      */
     public static function create($input): self
@@ -83,6 +94,11 @@ final class CreateGlobalSecondaryIndexAction
     public function getKeySchema(): array
     {
         return $this->keySchema;
+    }
+
+    public function getOnDemandThroughput(): ?OnDemandThroughput
+    {
+        return $this->onDemandThroughput;
     }
 
     public function getProjection(): Projection
@@ -116,6 +132,9 @@ final class CreateGlobalSecondaryIndexAction
         $payload['Projection'] = $v->requestBody();
         if (null !== $v = $this->provisionedThroughput) {
             $payload['ProvisionedThroughput'] = $v->requestBody();
+        }
+        if (null !== $v = $this->onDemandThroughput) {
+            $payload['OnDemandThroughput'] = $v->requestBody();
         }
 
         return $payload;